Human argument order in decorator
When you decorate a function with multiple @arg or @option like
```python
@arg('foo', help='Foo')
@arg('bar', help='Bar')
def do_some(options):
  print "The function!"

Then the order for the command is `cmd bar foo`. This is because
decorators are processed from in to out.
http://docs.python.org/reference/compound_stmts.html#function
This change could be harmful, because may break prior versions
which use positional arguments.